+++ tests/unit/processor/labeler/test_labeler_rule.py
@@ -214,3 +214,57 @@ class TestRule:
         document = {"applyrule": None}
 
         assert rule.matches(document)
+
+    def test_lucene_regex_matches_returns_true_for_matching_document(self):
+        rule_definition = {
+            "filter": 'applyrule: "/.*yes.*/"',
+            "labeler": {"label": {"reporter": ["windows"]}},
+        }
+        rule = LabelerRule._create_from_dict(rule_definition)
+        assert rule.matches({"applyrule": "yes"})
+        assert rule.matches({"applyrule": "yes!"})
+        assert rule.matches({"applyrule": "no? yes!"})
+
+    def test_lucene_regex_matches_returns_false_for_non_matching_document(self):
+        rule_definition = {
+            "filter": 'applyrule: "/.*yes.*/"',
+            "labeler": {"label": {"reporter": ["windows"]}},
+        }
+        rule = LabelerRule._create_from_dict(rule_definition)
+        non_matching_documents = [
+            {},
+            {"applyrule": "no"},
+            {"applyrule": "ye s"},
+            {"applyrule": "YES"},
+            {"wrong key": "yes"},
+        ]
+
+        for document in non_matching_documents:
+            assert not rule.matches(document)
+
+    def test_complex_lucene_regex_matches_returns_true_for_matching_document(self):
+        rule_definition = {
+            "filter": r'applyrule: "/(?:(?=.*[a-z])(?:(?=.*[A-Z])(?=.*[\d\W])|(?=.*\W)(?=.*\d))|(?=.*\W)(?=.*[A-Z])(?=.*\d)).{8,}/"',
+            # pylint: disable=line-too-long
+            "labeler": {"label": {"reporter": ["windows"]}},
+        }
+        rule = LabelerRule._create_from_dict(rule_definition)
+        assert rule.matches({"applyrule": "UPlo8888"})
+        assert rule.matches({"applyrule": "UPlo99999"})
+        assert rule.matches({"applyrule": "UPlo$$$$"})
+        assert rule.matches({"applyrule": "UP$$$$88"})
+
+    def test_complex_lucene_regex_does_not_match_returns_true_for_matching_document(self):
+        rule_definition = {
+            "filter": r'applyrule: "/(?:(?=.*[a-z])(?:(?=.*[A-Z])(?=.*[\d\W])|(?=.*\W)(?=.*\d))|(?=.*\W)(?=.*[A-Z])(?=.*\d)).{8,}/"',
+            # pylint: disable=line-too-long
+            "labeler": {"label": {"reporter": ["windows"]}},
+        }
+        rule = LabelerRule._create_from_dict(rule_definition)
+        assert not rule.matches({"applyrule": ""})
+        assert not rule.matches({"applyrule": "UPlo777"})
+        assert not rule.matches({"applyrule": "UP888888"})
+        assert not rule.matches({"applyrule": "lo888888"})
+        assert not rule.matches({"applyrule": "UPloXXXX"})
+        assert not rule.matches({"applyrule": "88888888"})
+        assert not rule.matches({"applyrule": "UPlo$$7"})